requirements.txt
aiogram
python-decouple
pika
aio_pika